create end point to return list of venues in a city with similar names
