+++ v2/server/schema/queries.js
@@ -16,6 +16,7 @@ export default gql`
       first: Int
     ): [Venue]
     similarVenues(id: ID!, radius: Int, first: Int): [Venue]
+    similarVenuesByName(name: String!, cityId: Int, first: Int): [Venue]
     cities(first: Int, query: String): [City]!
     me: User
   }
